# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Resonance API Documentation",
        "description": "API documentation for Resonance",
        "contact": {
            "email": "casier07@hotmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://resonance-be.ddev.site",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/artists": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Get all artists",
                "operationId": "2c8aed1387cca1c65d98da7ee11ac926",
                "parameters": [
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Field to sort by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "name",
                            "enum": [
                                "name",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "description": "Sort direction",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending_approval",
                                "verified",
                                "rejected"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in name and description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of artists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Artist"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 75
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string"
                                                },
                                                "last": {
                                                    "type": "string"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Artists"
                ],
                "summary": "Create a new artist",
                "operationId": "83d2888bdc6fac66b9a7cc81c4c950fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "David Guetta"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "French DJ and music producer known for electronic dance music.",
                                        "nullable": true
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Belgium",
                                        "nullable": true
                                    },
                                    "formed_year": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2002",
                                        "nullable": true
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "example": "https://resonance-be.ddev.site/storage/artists/david-guetta.jpg",
                                        "nullable": true
                                    },
                                    "source": {
                                        "type": "string",
                                        "enum": [
                                            "manual",
                                            "api"
                                        ],
                                        "example": "manual"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending_approval",
                                            "verified",
                                            "rejected"
                                        ],
                                        "example": "pending_approval"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Artist created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Artist"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/artists/{id}": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Get a specific artist",
                "operationId": "d0b68c06b8bc310fbe24c6d3ea80dac3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Artist UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artist details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Artist"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Artist not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Artists"
                ],
                "summary": "Update an artist",
                "operationId": "5a5632ea389f9a05336f8d035f5a44c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Artist UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "David Guetta"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "French DJ and music producer known for electronic dance music.",
                                        "nullable": true
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Belgium",
                                        "nullable": true
                                    },
                                    "formed_year": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2002",
                                        "nullable": true
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "example": "https://resonance-be.ddev.site/storage/artists/david-guetta.jpg",
                                        "nullable": true
                                    },
                                    "source": {
                                        "type": "string",
                                        "enum": [
                                            "manual",
                                            "api"
                                        ],
                                        "example": "manual"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending_approval",
                                            "verified",
                                            "rejected"
                                        ],
                                        "example": "verified"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Artist updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Artist"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Artist not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Artists"
                ],
                "summary": "Delete an artist",
                "operationId": "4f6d8bd2c8b4b738af7589b72348ac20",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Artist UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Artist deleted successfully"
                    },
                    "404": {
                        "description": "Artist not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Artist": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "name": {
                        "type": "string",
                        "example": "David Guetta"
                    },
                    "description": {
                        "type": "string",
                        "example": "French DJ and music producer known for electronic dance music.",
                        "nullable": true
                    },
                    "country": {
                        "type": "string",
                        "example": "Belgium",
                        "nullable": true
                    },
                    "formed_year": {
                        "type": "string",
                        "format": "date",
                        "example": "2002",
                        "nullable": true
                    },
                    "image_url": {
                        "type": "string",
                        "example": "https://resonance-be.ddev.site/storage/artists/david-guetta.jpg",
                        "nullable": true
                    },
                    "source": {
                        "type": "string",
                        "example": "manual"
                    },
                    "status": {
                        "type": "string",
                        "example": "verified"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}